+++ spec/hatchet/pip_spec.rb
@@ -15,11 +15,11 @@ RSpec.describe 'pip support' do
           remote: -----> Installing Python #{DEFAULT_PYTHON_FULL_VERSION}
           remote: -----> Installing pip #{PIP_VERSION}
           remote: -----> Installing dependencies using 'pip install -r requirements.txt'
-          remote:        Collecting typing-extensions==4.12.2 \\(from -r requirements.txt \\(line 5\\)\\)
-          remote:          Downloading typing_extensions-4.12.2-py3-none-any.whl.metadata \\(3.0 kB\\)
-          remote:        Downloading typing_extensions-4.12.2-py3-none-any.whl \\(37 kB\\)
+          remote:        Collecting typing-extensions==4.15.0 \\(from -r requirements.txt \\(line 5\\)\\)
+          remote:          Downloading typing_extensions-4.15.0-py3-none-any.whl.metadata \\(3.3 kB\\)
+          remote:        Downloading typing_extensions-4.15.0-py3-none-any.whl \\(44 kB\\)
           remote:        Installing collected packages: typing-extensions
-          remote:        Successfully installed typing-extensions-4.12.2
+          remote:        Successfully installed typing-extensions-4.15.0
           remote: -----> Running bin/post_compile hook
           remote:        BUILD_DIR=/tmp/build_.+
           remote:        CACHE_DIR=/tmp/codon/tmp/cache
@@ -57,18 +57,18 @@ RSpec.describe 'pip support' do
           remote: 
           remote: \\['',
           remote:  '/app',
-          remote:  '/app/.heroku/python/lib/python313.zip',
-          remote:  '/app/.heroku/python/lib/python3.13',
-          remote:  '/app/.heroku/python/lib/python3.13/lib-dynload',
-          remote:  '/app/.heroku/python/lib/python3.13/site-packages'\\]
+          remote:  '/app/.heroku/python/lib/python314.zip',
+          remote:  '/app/.heroku/python/lib/python3.14',
+          remote:  '/app/.heroku/python/lib/python3.14/lib-dynload',
+          remote:  '/app/.heroku/python/lib/python3.14/site-packages'\\]
           remote: 
-          remote: pip #{PIP_VERSION} from /app/.heroku/python/lib/python3.13/site-packages/pip \\(python 3.13\\)
+          remote: pip #{PIP_VERSION} from /app/.heroku/python/lib/python3.14/site-packages/pip \\(python 3.14\\)
           remote: Package           Version
           remote: ----------------- -------
           remote: pip               #{PIP_VERSION}
-          remote: typing_extensions 4.12.2
+          remote: typing_extensions 4.15.0
           remote: 
-          remote: <module 'typing_extensions' from '/app/.heroku/python/lib/python3.13/site-packages/typing_extensions.py'>
+          remote: <module 'typing_extensions' from '/app/.heroku/python/lib/python3.14/site-packages/typing_extensions.py'>
           remote: 
           remote: \\{
           remote:   "cache_restore_duration": [0-9.]+,
@@ -85,11 +85,11 @@ RSpec.describe 'pip support' do
           remote:   "pre_compile_hook": false,
           remote:   "python_install_duration": [0-9.]+,
           remote:   "python_version": "#{DEFAULT_PYTHON_FULL_VERSION}",
-          remote:   "python_version_major": "3.13",
+          remote:   "python_version_major": "3.14",
           remote:   "python_version_origin": ".python-version",
           remote:   "python_version_outdated": false,
           remote:   "python_version_pinned": false,
-          remote:   "python_version_requested": "3.13",
+          remote:   "python_version_requested": "3.14",
           remote:   "total_duration": [0-9.]+
           remote: \\}
         REGEX
@@ -103,7 +103,7 @@ RSpec.describe 'pip support' do
           remote: -----> Using cached install of Python #{DEFAULT_PYTHON_FULL_VERSION}
           remote: -----> Installing pip #{PIP_VERSION}
           remote: -----> Installing dependencies using 'pip install -r requirements.txt'
-          remote:        Requirement already satisfied: typing-extensions==4.12.2 \\(from -r requirements.txt \\(line 5\\)\\) \\(4.12.2\\)
+          remote:        Requirement already satisfied: typing-extensions==4.15.0 \\(from -r requirements.txt \\(line 5\\)\\) \\(4.15.0\\)
           remote: -----> Running bin/post_compile hook
           remote:        .+
           remote: -----> Saving cache
@@ -123,7 +123,7 @@ RSpec.describe 'pip support' do
           PYTHONPATH=/app
           PYTHONUNBUFFERED=true
           WEB_CONCURRENCY=2
-          pip #{PIP_VERSION} from /app/.heroku/python/lib/python3.13/site-packages/pip (python 3.13)
+          pip #{PIP_VERSION} from /app/.heroku/python/lib/python3.14/site-packages/pip (python 3.14)
         OUTPUT
         expect($CHILD_STATUS.exitstatus).to eq(0)
       end
@@ -149,14 +149,14 @@ RSpec.describe 'pip support' do
           remote: -----> Installing Python #{DEFAULT_PYTHON_FULL_VERSION}
           remote: -----> Installing pip #{PIP_VERSION}
           remote: -----> Installing dependencies using 'pip install -r requirements.txt'
-          remote:        Collecting typing-extensions==4.12.2 (from -r requirements.txt (line 5))
-          remote:          Downloading typing_extensions-4.12.2-py3-none-any.whl.metadata (3.0 kB)
+          remote:        Collecting typing-extensions==4.15.0 (from -r requirements.txt (line 5))
+          remote:          Downloading typing_extensions-4.15.0-py3-none-any.whl.metadata (3.3 kB)
           remote:        Collecting six==1.17.0 (from -r requirements.txt (line 6))
           remote:          Downloading six-1.17.0-py2.py3-none-any.whl.metadata (1.7 kB)
-          remote:        Downloading typing_extensions-4.12.2-py3-none-any.whl (37 kB)
+          remote:        Downloading typing_extensions-4.15.0-py3-none-any.whl (44 kB)
           remote:        Downloading six-1.17.0-py2.py3-none-any.whl (11 kB)
           remote:        Installing collected packages: typing-extensions, six
-          remote:        Successfully installed six-1.17.0 typing-extensions-4.12.2
+          remote:        Successfully installed six-1.17.0 typing-extensions-4.15.0
           remote: -----> Saving cache
         OUTPUT
       end
@@ -320,11 +320,11 @@ RSpec.describe 'pip support' do
           remote: -----> Installing Python 3.9.0
           remote: -----> Installing pip #{PIP_VERSION}, setuptools #{SETUPTOOLS_VERSION} and wheel #{WHEEL_VERSION}
           remote: -----> Installing dependencies using 'pip install -r requirements.txt'
-          remote:        Collecting typing-extensions==4.14.1 (from -r requirements.txt (line 2))
-          remote:          Downloading typing_extensions-4.14.1-py3-none-any.whl.metadata (3.0 kB)
-          remote:        Downloading typing_extensions-4.14.1-py3-none-any.whl (43 kB)
+          remote:        Collecting typing-extensions==4.15.0 (from -r requirements.txt (line 2))
+          remote:          Downloading typing_extensions-4.15.0-py3-none-any.whl.metadata (3.3 kB)
+          remote:        Downloading typing_extensions-4.15.0-py3-none-any.whl (44 kB)
           remote:        Installing collected packages: typing-extensions
-          remote:        Successfully installed typing-extensions-4.14.1
+          remote:        Successfully installed typing-extensions-4.15.0
           remote: -----> Saving cache
         OUTPUT
         app.commit!
@@ -334,7 +334,7 @@ RSpec.describe 'pip support' do
         # regex variant can be removed once support for Python 3.10 and older is dropped.
         expect(clean_output(app.output)).to include(<<~OUTPUT)
           remote: -----> Installing dependencies using 'pip install -r requirements.txt'
-          remote:        Requirement already satisfied: typing-extensions==4.14.1 (from -r requirements.txt (line 2)) (4.14.1)
+          remote:        Requirement already satisfied: typing-extensions==4.15.0 (from -r requirements.txt (line 2)) (4.15.0)
           remote: -----> Saving cache
         OUTPUT
       end